+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add RMON event configuration to `nxos_snmp` resource and data source
 - Add management interface configuration to `nxos_system` resource and data source
 - Add LLDP configuration to `nxos_system` resource and data source
 - Add platform entity configuration to `nxos_system` resource and data source